% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-basis.R
\name{calc_omega_norm}
\alias{calc_omega_norm}
\title{Calulate basis function normalizing constant}
\usage{
calc_omega_norm(W, Q_inv)
}
\arguments{
\item{W}{The sparse basis function matrix}

\item{Q_inv}{The inverse of the sparse precision matrix}
}
\value{
A set of normalizing weights for the basis functions
}
\description{
Calculate the basis function normalizing constant so that all basis functions have the same marginal variance
}
