% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{check_corr_fun}
\alias{check_corr_fun}
\title{Check if the correlation function type is valid}
\usage{
check_corr_fun(corr_fun)
}
\arguments{
\item{corr_fun}{is the correlation function}
}
\description{
this function checks if the correlation function type is valid
}
\keyword{internal}
