% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{code_test}
\alias{code_test}
\title{Small temperature data from simulation for testing}
\format{
A data frame with 10000 rows and 4 variables:
\describe{
\item{Lon}{The longitude of the obseravtion}
\item{Lat}{The latitude of the obseravtion}
\item{MaskTemp}{The simulated temperature observations with a mask that is used for training the model}
\item{TrueTemp}{The simulated temperature observations without a mask that is used for evaluating the model}
}
}
\source{
\href{https://github.com/finnlindgren/heatoncomparison}{GitHub}
}
\usage{
data("code_test")
}
\description{
A dataset containing the temperature measurements and locations
from simulated data used in Heaton et al, A Case Study
Competition Among Methods for Analyzing Large Spatial Data,
Journal of Agricultural, Biological and Environmental Statistics,
2018: http://link.springer.com/article/10.1007/s13253-018-00348-w
}
\examples{
library(BayesMRA)
library(tidyverse)
data("code_test")
ggplot(code_test, aes(x = Lon, y = Lat, fill = TrueTemp)) +
 geom_raster()
}
\references{
Heaton et al, A Case Study
Competition Among Methods for Analyzing Large Spatial Data,
Journal of Agricultural, Biological and Environmental Statistics,
2018:

(\href{http://link.springer.com/article/10.1007/s13253-018-00348-w}{JABES})
}
\keyword{datasets}
