% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvn_smw.R
\name{dmvn_smw}
\alias{dmvn_smw}
\title{dmvn_smw_mra}
\usage{
dmvn_smw(
  y,
  X,
  beta,
  tW,
  tWW,
  Q_alpha_tau2,
  sigma2,
  Rstruct = NULL,
  logd = TRUE
)
}
\arguments{
\item{y}{is a \eqn{n}{n} vector of Gaussian data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of fixed effects (like latitude, elevation, etc)}

\item{beta}{is the regression parameter beta}

\item{tW}{is the transpose of the sparse Wendland basis matrix of class spam}

\item{tWW}{is the transpose of the sparse Wendland basis matrix multiplied by itself of class spam}

\item{Q_alpha_tau2}{is the prior precision matrix for random effects alpha}

\item{sigma2}{is the residual error}

\item{Rstruct}{is the Cholesky prior precision matrix for random effects alpha}

\item{logd}{is a logical value of whether to calculate the log density (\code{logd = TRUE}) or the density on the data scale (\code{logd = FALSE})}
}
\value{
The (log) density of a normal distribution with mean \eqn{\mathbf{X} \boldsymbol{\beta}} and covariance matrix \eqn{\sigma^2 \mathbf{I} + \mathbf{W} \mathbf{Q}_{\alpha_{\tau^2}} \mathbf{W}'}
}
\description{
dmvn_smw_mra
}
