% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess}
\alias{ess}
\title{Code for the elliptical slice sampler}
\usage{
ess(
  current,
  prior,
  prior_mean,
  pars,
  log_like_fun,
  update_pars_fun,
  max_iter = 500
)
}
\arguments{
\item{current}{The current state of the parameter of interest in the
MCMC algorithm.}

\item{prior}{A sample from the mean zero, Gaussian prior of the
parameter of interest.}

\item{prior_mean}{The mean, if nonzero, of the Gaussian prior
distribution for the parameter of interest.}

\item{pars}{The set of parameters needed to evaluation the
log-likelihood function for the parameter of interest. At a minimum,
the pars object must contain the variables \code{parameter} which is
the parameter to be updated, \code{par_name} which is a string name
of the variable to update, \code{verbose} which is a boolean option for verbose output,
and the ess performance tracking variables \code{num_calls_ess},
\code{num_max_contract}, and \code{num_unusual_contract}.}

\item{log_like_fun}{The log-likelihood function for the parameter of
interest.}

\item{update_pars_fun}{The function to update the parameters
\code{pars}.}

\item{max_iter}{The maximum number of iterations to perform for the
elliptical slice sampler.}
}
\value{
An updated set of \code{pars} after an elliptical slice sampling update
}
\description{
Code for the elliptical slice sampler
}
