% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_integer}
\alias{is_integer}
\title{Check if value is an integer or integer-like}
\usage{
is_integer(x, n)
}
\arguments{
\item{x}{is the input}

\item{n}{is the number of inputs}
}
\description{
this function checks if the input is an integer scalar (integer-like value -- i.e., both 1L and 1.0 pass this check)
}
\keyword{internal}
