% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_integer_matrix}
\alias{is_integer_matrix}
\title{Check if an integer matrix of dimension \eqn{n \times m}{n x m}}
\usage{
is_integer_matrix(x, n, m)
}
\arguments{
\item{x}{is the input}

\item{n}{is the input matrix rows}

\item{m}{is the input matrix columns}
}
\description{
this function checks if the input is an integer matrix of dimension \eqn{n \times m}{n x m}
}
\keyword{internal}
