% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_numeric_vector}
\alias{is_numeric_vector}
\title{Check if a numeric vector of length n}
\usage{
is_numeric_vector(x, n)
}
\arguments{
\item{x}{is the input}

\item{n}{is the input vector length}
}
\description{
this function checks if the input is a numeric (double) vector
}
\keyword{internal}
