% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_numeric_with_na}
\alias{is_numeric_with_na}
\title{Check if a numeric value with NA values allowed}
\usage{
is_numeric_with_na(x, n)
}
\arguments{
\item{x}{is the input}

\item{n}{is the input length}
}
\description{
this function checks if the input is a scalar (double) with NA values allowed
}
\keyword{internal}
