% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_positive_numeric}
\alias{is_positive_numeric}
\title{Check if a positive numeric value}
\usage{
is_positive_numeric(x, n)
}
\arguments{
\item{x}{is the input}

\item{n}{is the input length}
}
\description{
this function checks if the input is a positive scalar (double)
}
\keyword{internal}
