% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-Q.R
\name{make_Q}
\alias{make_Q}
\title{Generate CAR precision matrix}
\usage{
make_Q(n, phi)
}
\arguments{
\item{n}{is the length of the coefficient vector for the penalized bspline model}

\item{phi}{is a number between -1 and 1 that defines the strength of the  autoregressive process. Typically this will be set to 1 for use as a prior in Bayesian penalized splines}
}
\value{
an \eqn{n \times n}{n x n} sparse Matrix of class "dgCMatrix" (see Matrix package for details)
}
\description{
A function for setting up a conditional autoregressive (CAR) precision matrix for use as a prior in Bayesian penalized splines
}
