% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_mra.R
\name{mcmc_mra}
\alias{mcmc_mra}
\title{Bayesian Polya-gamma regression}
\usage{
mcmc_mra(
  y,
  X,
  locs,
  params,
  priors,
  M = 4,
  n_coarse_grid = 10,
  n_cores = 1L,
  inits = NULL,
  config = NULL,
  verbose = FALSE,
  use_spam = TRUE,
  n_chain = 1
)
}
\arguments{
\item{y}{is a \eqn{n}{n} vector of Gaussian data.}

\item{X}{is a \eqn{n \times p}{n x p} matrix of fixed effects (like latitude, elevation, etc)}

\item{locs}{is a \eqn{n \times 2}{n x 2} matrix of observation locations.}

\item{params}{is a list of parameter settings. The list
\code{params} must contain the following values:
\itemize{
\item \code{n_adapt}: A positive integer number of adaptive MCMC iterations.
\item \code{n_mcmc}: A positive integer number of total MCMC iterations
post adaptation.
\item \code{n_thin}: A positive integer number of MCMC iterations per saved
sample.
\item \code{n_message}: A positive integer number of frequency of iterations
to output a progress message. For example, \code{n_message = 50}
outputs progress messages every 50 iterations.
}}

\item{priors}{is the list of prior settings.}

\item{n_cores}{is the number of cores for parallel computation using openMP.}

\item{inits}{is the list of initial values if the user wishes to specify initial values. If these values are not specified, then the initial values will be randomly sampled from the prior.}

\item{config}{is the list of configuration values if the user wishes to specify initial values. If these values are not specified, then default a configuration will be used.}

\item{n_chain}{is the MCMC chain id. The default is 1.}
}
\description{
this function runs the Bayesian multinomial regression using Polya-gamma data augmentation
}
