% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_observations.R
\name{partition_observations}
\alias{partition_observations}
\title{Partition observations into MRA grid}
\usage{
partition_observations(locs, M)
}
\arguments{
\item{locs}{is a \eqn{n \times 2}{n x 2} matrix of observation locations.}

\item{M}{The number of resolutions.}
}
\value{
A data.frame of partition ids for each location
}
\description{
Partition the observation locations into MRA grids for use in \code{\link[=mcmc_mra_vecchia]{mcmc_mra_vecchia()}}
}
