% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MRA_grid.R
\name{plot_MRA_grid}
\alias{plot_MRA_grid}
\title{Plot the MRA spatial grid}
\usage{
plot_MRA_grid(MRA, base_size = 12, file = NULL, width = 16, height = 9)
}
\arguments{
\item{MRA}{The output from \code{mra_wendland_2d()}}

\item{base_size}{The base size for the plot}

\item{file}{If \code{file = NULL}, the ggplot object is returned. If \code{file} is not NULL, an image is saved to the file path specified by \code{file}}

\item{width}{If a file path is specified, \code{width} determines the width of the saved image (in inches)}

\item{height}{If a file path is specified, \code{height} determines the height of the saved image (in inches)}
}
\value{
Either a ggplot object (if \code{file = NULL}) or a saved image file with no return (\code{file} is not NULL)
}
\description{
Plot the MRA  spatial grid from the output of \code{mra_wendland_2d()}
}
