% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_vs_fitted.R
\name{plot_sim_vs_fitted}
\alias{plot_sim_vs_fitted}
\title{Plot the fitted model parameters vs. simulated model parameters}
\usage{
plot_sim_vs_fitted(
  out,
  X,
  beta,
  W,
  alpha,
  base_size = 12,
  file = NULL,
  width = 16,
  height = 9
)
}
\arguments{
\item{out}{The output from \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}

\item{X}{The simulated regression covariates}

\item{beta}{The simulated regression coefficients}

\item{W}{The simulated spatial basis functions}

\item{alpha}{The simulated spatial basis coefficients}

\item{base_size}{The base size for the plot}

\item{file}{If \code{file = NULL}, the ggplot object is returned. If \code{file} is not NULL, an image is saved to the file path specified by \code{file}}

\item{width}{If a file path is specified, \code{width} determines the width of the saved image (in inches)}

\item{height}{If a file path is specified, \code{height} determines the height of the saved image (in inches)}
}
\value{
Either a ggplot object of the fitted vs. simulated parameter values (if \code{file = NULL}) or a saved image file with no return (\code{file} is not NULL)
}
\description{
Plot the fitted model parameters vs. simulated model parameters
}
