% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sparse_matrix.R
\name{plot_sparse_matrix}
\alias{plot_sparse_matrix}
\title{Plot the sparsity structure of a spam matrix object using ggplot2}
\usage{
plot_sparse_matrix(x, tile_size = 1, base_size = 12, title = "")
}
\arguments{
\item{x}{The sparse matrix which must be of class "spam"}

\item{tile_size}{The size of the ggplot2 tile}

\item{base_size}{The base size for the plot}

\item{title}{The title for the plot}
}
\value{
A ggplot of the sparse matrix structure of a matrix of class "spam"
}
\description{
Plot the sparsity of a spam matrix object using ggplot2
}
