% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test_data.R
\name{plot_test_data}
\alias{plot_test_data}
\title{Plot the observed and "true" (holdout) spatial data}
\usage{
plot_test_data(data, base_size = 12, file = NULL, width = 16, height = 9, ...)
}
\arguments{
\item{data}{A dataframe with four columns: Lat, Lon, Observed, Fitted}

\item{base_size}{The base size for the plot}

\item{...}{Other ggplot2 options}
}
\value{

}
\description{
Plot the observed and "true" (holdout) spatial data
}
