% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace_alpha.R
\name{plot_trace_alpha}
\alias{plot_trace_alpha}
\title{Trace plots from the output of \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}
\usage{
plot_trace_alpha(
  out,
  base_size = 12,
  alpha = 0.1,
  n_per_res = 10,
  file = NULL,
  width = 16,
  height = 9
)
}
\arguments{
\item{out}{The output from \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}

\item{base_size}{The base size for the plot}

\item{alpha}{The plot transparency}

\item{n_per_res}{The number of alpha parameters to sample per resolution}

\item{file}{If \code{file = NULL}, the ggplot object is returned. If \code{file} is not NULL, an image is saved to the file path specified by \code{file}}

\item{width}{If a file path is specified, \code{width} determines the width of the saved image (in inches)}

\item{height}{If a file path is specified, \code{height} determines the height of the saved image (in inches)}
}
\value{
Either a ggplot object of the trace plots for a subset of the spatial coefficients alpha (if \code{file = NULL}) or a saved image file with no return (\code{file} is not NULL)
}
\description{
Trace plots from the output of \code{mcmc_mra()} or \code{mcmc_mra_integrated()}
}
