% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mra.R
\name{predict_mra}
\alias{predict_mra}
\title{Predict for new data using the MRA model}
\usage{
predict_mra(out, new_data)
}
\arguments{
\item{out}{The fitted mra object from \code{mcmc_mra()}}

\item{new_data}{A list of new data at predicted locations with two inputs: \code{locs_pred} and \code{X_pred} representing the prediction locations and covariates, respectively}
}
\value{
A list of prediction objects: \code{MRA_pred} is the MRA grid information for the predicted locations. \code{Xbeta_pred} is the predicted fixed effects, \code{Walpha_pred} is the predicted spatial process, \code{y_pred} is the predicted response
}
\description{
Predict for new data using the MRA model
}
