% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-pg_mvgp_univariate_mra.R
\name{predict_pg_mvgp_univariate_mra}
\alias{predict_pg_mvgp_univariate_mra}
\title{Bayesian Polya-gamma regression prediction}
\usage{
predict_pg_mvgp_univariate_mra(
  out,
  X,
  X_pred,
  locs,
  locs_pred,
  n_cores = 1L,
  progress = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{out}{is a list of MCMC outputs from pgSPLM}

\item{X}{is a \eqn{n \times p}{n x p} matrix of covariates at the observed locations.}

\item{X_pred}{is a \eqn{n_{pred} \times p}{n_{pred} x p} matrix of covariates at the locations where predictions are to be made.}

\item{locs}{is a \eqn{n \times 2}{n x 2} matrix of locations where observations were taken.}

\item{locs_pred}{is a \eqn{n_pred \times 2}{n_pred x 2} matrix of locations where predictions are to be made.}

\item{n_cores}{is the number of cores for parallel computation using openMP.}

\item{progress}{is a logicial input that determines whether to print a progress bar.}

\item{verbose}{is a logicial input that determines whether to print more detailed messages.}
}
\description{
this function generates predictions from the Bayesian multinomial regression using Polya-gamma data augmentation
}
