% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvn_arma_scalar}
\alias{rmvn_arma_scalar}
\title{A function for sampling from conditional multivariate normal distributions with mean A^{-1}b and covariance matrix A^{-1}.}
\usage{
rmvn_arma_scalar(a, b)
}
\arguments{
\item{a}{\code{a} A scalar for the Gaussian full conditional distribution precision.}

\item{b}{\code{b} A \eqn{d} \code{vector} for the Gaussian full conditional distribution mean.}
}
\description{
A function for sampling from conditional multivariate normal distributions with mean A^{-1}b and covariance matrix A^{-1}.
}
\examples{
set.seed(111)
a <- 4
b <- rnorm(1)
sample <- rmvn_arma_scalar(a, b)

}
