% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale_estimates.R
\name{unscale_beta}
\alias{unscale_beta}
\title{Transform the beta parameters to restore the parameters to the original scale.
The functions \code{mcmc_mra()} and \code{mcmc_mra_integrated()} fit the regression
coefficients using centered and scaled response and covariate values. This
function undoes that transformation to scale the parameter estimates to
their usual (uncentered and unscaled) scale
estimates}
\usage{
unscale_beta(out)
}
\arguments{
\item{out}{The output of \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}
}
\value{
A matrix of the size of parameters in out$beta of transformed
regression parameters equivalent to fitting the model on the original scale
}
\description{
Transform the beta parameters to restore the parameters to the original scale.
The functions \code{mcmc_mra()} and \code{mcmc_mra_integrated()} fit the regression
coefficients using centered and scaled response and covariate values. This
function undoes that transformation to scale the parameter estimates to
their usual (uncentered and unscaled) scale
estimates
}
