% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-tuning.R
\name{update_tuning}
\alias{update_tuning}
\title{this function updates the univariate Gaussian random walk proposal Metropolis-Hastings tuning parameters}
\usage{
update_tuning(k, accept, tune)
}
\arguments{
\item{k}{is a positive integer that is the current MCMC iteration.}

\item{accept}{is a number between 0 and 1 that represents the batch accpetance rate. The target of this adaptive tuning algorithm is an acceptance rate of 0.44.}

\item{tune}{is a positive number that is the univariate Gaussian random walk proposal standard deviation.}
}
\description{
this function updates the univariate Gaussian random walk proposal Metropolis-Hastings tuning parameters
}
\keyword{internal}
