% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-tuning.R
\name{update_tuning_mv_mat}
\alias{update_tuning_mv_mat}
\title{this function updates multiple block Gaussian random walk proposal Metropolis-Hastings tuning parameters}
\usage{
update_tuning_mv_mat(
  k,
  accept,
  lambda,
  batch_samples,
  Sigma_tune,
  Sigma_tune_chol
)
}
\arguments{
\item{k}{is a positive integer that is the current MCMC iteration.}

\item{accept}{is a \eqn{p}{p}-dimensional vector of positive number between 0 and 1 that represents the batch accpetance rate. The target of this adaptive tuning algorithm is an acceptance rate of between 0.44 (a univariate update) and 0.234 (5+ dimensional upate).}

\item{lambda}{is a is a \eqn{p}{p}-dimensional vector of positive scalars that scales the covariance matrix Sigma_tune and diminishes towards 0 as k increases.}

\item{batch_samples}{is a \eqn{50 \times d \times p}{50 x d x p} dimensional array that consists of the 50 batch samples of the d-dimensional for each of the \code{p}{p} parameter groups being sampled.}

\item{Sigma_tune}{is a \eqn{d \times d \times p}{d x d x p} array of \eqn{d \times d}{d x d} positive definite covariance matrix of the batch samples used to generate the multivariate Gaussian random walk proposal.}

\item{Sigma_tune_chol}{is the \eqn{d \times d \times p}{d x d x p} array of \eqn{d \times d}{d x d} Cholesky decompositions of Sigma_tune.}
}
\description{
this function updates multiple block Gaussian random walk proposal Metropolis-Hastings tuning parameters
}
\keyword{internal}
