% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-tuning.R
\name{update_tuning_vec}
\alias{update_tuning_vec}
\title{this function updates a vector block of univariate Gaussian random walk proposal Metropolis-Hastings tuning parameters}
\usage{
update_tuning_vec(k, accept, tune)
}
\arguments{
\item{k}{is a positive integer that is the current MCMC iteration.}

\item{accept}{is a vector of positive numbers between 0 and 1 that represents the batch accpetance rate. The target of this adaptive tuning algorithm is an acceptance rate of 0.44.}

\item{tune}{is a vector of positive numbers that are the univariate Gaussian random walk proposal standard deviation.}
}
\description{
this function updates a vector block of univariate Gaussian random walk proposal Metropolis-Hastings tuning parameters
}
\keyword{internal}
