% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_sympd_matrix}
\alias{is_sympd_matrix}
\title{Check if a symmetric positive definite numeric matrix of dimension \eqn{n \times n}{n x n}}
\usage{
is_sympd_matrix(x, n)
}
\arguments{
\item{x}{is the input}

\item{n}{is the input matrix dimension (assuming a square matrix)}
}
\description{
this function checks if the input is a symmetric positive definite matrix
}
\keyword{internal}
