% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-Q-alpha_2d.R
\name{make_Q_alpha_2d}
\alias{make_Q_alpha_2d}
\title{Generate CAR precision matrix}
\usage{
make_Q_alpha_2d(n_dims, phi, use_spam = TRUE, prec_model = "CAR")
}
\arguments{
\item{phi}{is a number between -1 and 1 that defines the strength of the  autoregressive process. Typically this will be set to 1 for use as a prior in Bayesian penalized splines}

\item{use_spam}{is a boolean flag to determine whether the output is a list of spam matrix objects (\code{use_spam = TRUE}) or a an \eqn{n \times n}{n x n} sparse Matrix of class "dgCMatrix" \code{use_spam = FALSE}(see Matrix package for details)}

\item{n}{is the length of the coefficient vector for the penalized bspline model}
}
\value{
a list of  \eqn{n \times n}{n x n} sparse spam matrices or Matrix matrices of class "dgCMatrix" (see Matrix package for details)
}
\description{
A function for setting up a conditional autoregressive (CAR) precision matrix for use as a prior in Bayesian penalized splines
}
