% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-Q-alpha_tau2.R
\name{make_Q_alpha_tau2}
\alias{make_Q_alpha_tau2}
\title{Title}
\usage{
make_Q_alpha_tau2(Q_alpha, tau2, use_spam = TRUE)
}
\arguments{
\item{Q_alpha}{a list of length M composed of matrices that are the correlation structure of the CAR prior on beta}

\item{tau2}{a vector of length M that contains the CAR prior precisions}

\item{use_spam}{a boolean that determines if the output matrix is of class "spam" (\code{use_spam = TRUE}) or of class "dgCMatrix" (\code{use_spam = FALSE}; see Matrix package for details)}
}
\value{

}
\description{
Title
}
