% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra-wendland-2d.R
\name{mra_wendland_2d}
\alias{mra_wendland_2d}
\title{Code to construct the mutli-resolution sparse basis function representation for fitting spatial processes}
\usage{
mra_wendland_2d(
  locs,
  M = 4,
  n_neighbors = 68,
  n_coarse_grid = 10,
  n_padding = 5L,
  use_spam = TRUE
)
}
\arguments{
\item{locs}{The location variables in 2 dimensions over which to construct the basis function representation}

\item{M}{The number of resolutions}

\item{n_neighbors}{}

\item{use_spam}{}

\item{padding}{}
}
\value{

}
\description{
Code to construct the mutli-resolution sparse basis function representation for fitting spatial processes
}
