% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted_MRA.R
\name{plot_fitted_MRA}
\alias{plot_fitted_MRA}
\title{Plot the fitted MRA process}
\usage{
plot_fitted_MRA(
  out,
  preds,
  base_size = 12,
  file = NULL,
  width = 16,
  height = 9,
  ...
)
}
\arguments{
\item{out}{The output from \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}

\item{base_size}{The base size for the plot}

\item{...}{}

\item{data}{A dataframe with four columns: Lat, Lon, Observed, Fitted}
}
\value{

}
\description{
Plot the fitted spatial process from the output of \code{mcmc_mra()} or \code{mcmc_mra_integrated()} for each resolution.
}
