% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_params.R
\name{plot_posterior_params}
\alias{plot_posterior_params}
\title{Posterior distribution plots from the output of \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}
\usage{
plot_posterior_params(out, base_size = 12, alpha = 0.1, ...)
}
\arguments{
\item{out}{The output from \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}

\item{base_size}{The base size for the plot}

\item{alpha}{The plot transparency}
}
\value{
Plots of the posterior distribution of model parameters
}
\description{
Posterior distribution plots from the output of \code{mcmc_mra()} or \code{mcmc_mra_integrated()}
}
