% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_vs_fitted.R
\name{plot_sim_vs_fitted}
\alias{plot_sim_vs_fitted}
\title{Plot the fitted model parameters vs. simulated model parameters}
\usage{
plot_sim_vs_fitted(
  out,
  X,
  beta,
  W,
  alpha,
  base_size = 12,
  file = NULL,
  width = 16,
  height = 9,
  ...
)
}
\arguments{
\item{out}{The output from \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}

\item{X}{The simulated regression covariates}

\item{beta}{The simulated regression coefficients}

\item{W}{The simulated spatial basis functions}

\item{alpha}{The simulated spatial basis coefficients}

\item{base_size}{The base size for the plot}

\item{tile_size}{The size of the ggplot2 tile}

\item{title}{The title for the plot}
}
\value{

}
\description{
Plot the fitted model parameters vs. simulated model parameters
}
