% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Trace plots from the output of \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}
\usage{
plot_trace(out, base_size = 12)
}
\arguments{
\item{out}{The output from \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}

\item{base_size}{The base size for the plot}
}
\value{
Trace plots of the parameters
}
\description{
Trace plots from the output of \code{mcmc_mra()} or \code{mcmc_mra_integrated()}
}
