% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace_alpha.R
\name{plot_trace_alpha}
\alias{plot_trace_alpha}
\title{Trace plots from the output of \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}
\usage{
plot_trace_alpha(out, base_size = 12, alpha = 0.1, n_per_res = 10, ...)
}
\arguments{
\item{out}{The output from \code{mcmc_mra()} or \code{mcmc_mra_integrated()}}

\item{base_size}{The base size for the plot}

\item{alpha}{The plot transparency}

\item{n_per_res}{The number of alpha parameters to sample per resolutoin}
}
\value{
Trace plots of the alpha parameter
}
\description{
Trace plots from the output of \code{mcmc_mra()} or \code{mcmc_mra_integrated()}
}
