% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale_estimates.R
\name{unscale_sigma2}
\alias{unscale_sigma2}
\title{Transform the sigma2 parameter to restore the parameter to the original scale.
The function \code{mcmc_mra()} fits the regression variance using a centered and
scaled response. This function undoes that transformation to scale the
parameter estimate to its usual (uncentered and unscaled) scale}
\usage{
unscale_sigma2(out)
}
\arguments{
\item{out}{The output of \code{mcmc_mra()}}
}
\value{
A vector of the size of out$sigma2 of transformed regression
variance equivalent to fitting the model on the original scale
}
\description{
Transform the sigma2 parameter to restore the parameter to the original scale.
The function \code{mcmc_mra()} fits the regression variance using a centered and
scaled response. This function undoes that transformation to scale the
parameter estimate to its usual (uncentered and unscaled) scale
}
